perf: v4.4.6-1 (#364)
