+++ projects/app/public/locales/en/common.json
@@ -105,6 +105,7 @@
   },
   "common": {
     "Add": "Add",
+    "Close": "Clow",
     "Collect": "Collect",
     "Copy": "Copy",
     "Copy Successful": "Copy Successful",
