+++ projects/app/public/locales/zh/common.json
@@ -105,6 +105,7 @@
   },
   "common": {
     "Add": "添加",
+    "Close": "关闭",
     "Collect": "收藏",
     "Copy": "复制",
     "Copy Successful": "复制成功",
