+++ projects/app/src/components/ChatBox/index.tsx
@@ -1171,9 +1171,8 @@ export const useChatBox = () => {
         const historyDom = document.getElementById('history');
         if (!historyDom) return;
         const dom = Array.from(historyDom.children).map((child, i) => {
-          const avatar = `<img src="${
-            child.querySelector<HTMLImageElement>('.avatar')?.src
-          }" alt="" />`;
+          const avatar = `<img src="${child.querySelector<HTMLImageElement>('.avatar')
+            ?.src}" alt="" />`;
 
           const chatContent = child.querySelector<HTMLDivElement>('.markdown');
