+++ projects/app/src/pages/api/core/dataset/searchTest.ts
@@ -38,8 +38,8 @@ export default withNextCors(async function handler(req: NextApiRequest, res: Nex
         select id, q, a, source, file_id, (vector <#> '[${
           vectors[0]
         }]') * -1 AS score from ${PgDatasetTableName} where kb_id='${kbId}' AND user_id='${userId}' order by vector <#> '[${
-        vectors[0]
-      }]' limit 12;
+          vectors[0]
+        }]' limit 12;
         COMMIT;`
     );
