+++ projects/app/src/pages/app/detail/components/AdEdit/components/Nodes/NodeAnswer.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import { NodeProps } from 'reactflow';
 import NodeCard from '../modules/NodeCard';
-import { FlowModuleItemType } from '@/types/flow';
+import { FlowModuleItemType } from '@/types/core/app/flow';
 import Container from '../modules/Container';
 import RenderInput from '../render/RenderInput';
 import RenderOutput from '../render/RenderOutput';
 
 const NodeAnswer = ({ data }: NodeProps<FlowModuleItemType>) => {
-  const { moduleId, inputs, outputs, onChangeNode } = data;
+  const { moduleId, inputs, outputs } = data;
   return (
     <NodeCard minW={'400px'} {...data}>
       <Container borderTop={'2px solid'} borderTopColor={'myGray.200'}>
-        <RenderInput moduleId={moduleId} onChangeNode={onChangeNode} flowInputList={inputs} />
-        <RenderOutput onChangeNode={onChangeNode} moduleId={moduleId} flowOutputList={outputs} />
+        <RenderInput moduleId={moduleId} flowInputList={inputs} />
+        <RenderOutput moduleId={moduleId} flowOutputList={outputs} />
       </Container>
     </NodeCard>
   );
