+++ projects/app/src/pages/app/detail/components/AdEdit/components/Nodes/NodeEmpty.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NodeProps } from 'reactflow';
 import NodeCard from '../modules/NodeCard';
-import { FlowModuleItemType } from '@/types/flow';
+import { FlowModuleItemType } from '@/types/core/app/flow';
 
 const NodeAnswer = ({ data }: NodeProps<FlowModuleItemType>) => {
   return <NodeCard {...data}></NodeCard>;
