+++ projects/app/src/pages/app/detail/components/AdEdit/index.tsx
@@ -1,89 +1,45 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react';
-import ReactFlow, {
-  Background,
-  Controls,
-  ReactFlowProvider,
-  addEdge,
-  useNodesState,
-  useEdgesState,
-  XYPosition,
-  Connection,
-  useViewport
-} from 'reactflow';
+import ReactFlow, { Background, Controls, ReactFlowProvider } from 'reactflow';
 import { Box, Flex, IconButton, useTheme, useDisclosure } from '@chakra-ui/react';
 import { SmallCloseIcon } from '@chakra-ui/icons';
 import {
   edgeOptions,
   connectionLineStyle,
   FlowModuleTypeEnum,
-  FlowInputItemTypeEnum,
-  FlowValueTypeEnum
+  FlowInputItemTypeEnum
 } from '@/constants/flow';
-import { appModule2FlowNode, appModule2FlowEdge } from '@/utils/adapt';
-import {
-  FlowModuleItemType,
-  FlowModuleTemplateType,
-  FlowOutputTargetItemType,
-  type FlowModuleItemChangeProps
-} from '@/types/flow';
+import { FlowOutputTargetItemType } from '@/types/core/app/flow';
 import { AppModuleItemType } from '@/types/app';
-import { customAlphabet } from 'nanoid';
 import { useRequest } from '@/hooks/useRequest';
 import type { AppSchema } from '@/types/mongoSchema';
 import { useUserStore } from '@/store/user';
-import { useToast } from '@/hooks/useToast';
 import { useTranslation } from 'next-i18next';
 import { useCopyData } from '@/hooks/useCopyData';
 import dynamic from 'next/dynamic';
+import styles from './index.module.scss';
+import { AppTypeEnum } from '@/constants/app';
 
 import MyIcon from '@/components/Icon';
 import ButtonEdge from './components/modules/ButtonEdge';
 import MyTooltip from '@/components/MyTooltip';
 import TemplateList from './components/TemplateList';
 import ChatTest, { type ChatTestComponentRef } from './components/ChatTest';
+import FlowProvider, { useFlowStore } from './components/Provider';
 
-const ImportSettings = dynamic(() => import('./components/ImportSettings'), {
-  ssr: false
-});
-const NodeChat = dynamic(() => import('./components/Nodes/NodeChat'), {
-  ssr: false
-});
-const NodeKbSearch = dynamic(() => import('./components/Nodes/NodeKbSearch'), {
-  ssr: false
-});
-const NodeHistory = dynamic(() => import('./components/Nodes/NodeHistory'), {
-  ssr: false
-});
-const NodeTFSwitch = dynamic(() => import('./components/Nodes/NodeTFSwitch'), {
-  ssr: false
-});
-const NodeAnswer = dynamic(() => import('./components/Nodes/NodeAnswer'), {
-  ssr: false
-});
-const NodeQuestionInput = dynamic(() => import('./components/Nodes/NodeQuestionInput'), {
-  ssr: false
-});
-const NodeCQNode = dynamic(() => import('./components/Nodes/NodeCQNode'), {
-  ssr: false
-});
-const NodeVariable = dynamic(() => import('./components/Nodes/NodeVariable'), {
-  ssr: false
-});
-const NodeUserGuide = dynamic(() => import('./components/Nodes/NodeUserGuide'), {
-  ssr: false
-});
-const NodeExtract = dynamic(() => import('./components/Nodes/NodeExtract'), {
-  ssr: false
-});
-const NodeHttp = dynamic(() => import('./components/Nodes/NodeHttp'), {
-  ssr: false
-});
+const ImportSettings = dynamic(() => import('./components/ImportSettings'));
+const NodeChat = dynamic(() => import('./components/Nodes/NodeChat'));
+const NodeKbSearch = dynamic(() => import('./components/Nodes/NodeKbSearch'));
+const NodeHistory = dynamic(() => import('./components/Nodes/NodeHistory'));
+const NodeTFSwitch = dynamic(() => import('./components/Nodes/NodeTFSwitch'));
+const NodeAnswer = dynamic(() => import('./components/Nodes/NodeAnswer'));
+const NodeQuestionInput = dynamic(() => import('./components/Nodes/NodeQuestionInput'));
+const NodeCQNode = dynamic(() => import('./components/Nodes/NodeCQNode'));
+const NodeVariable = dynamic(() => import('./components/Nodes/NodeVariable'));
+const NodeUserGuide = dynamic(() => import('./components/Nodes/NodeUserGuide'));
+const NodeExtract = dynamic(() => import('./components/Nodes/NodeExtract'));
+const NodeHttp = dynamic(() => import('./components/Nodes/NodeHttp'));
 
 import 'reactflow/dist/style.css';
-import styles from './index.module.scss';
-import { AppTypeEnum } from '@/constants/app';
-
-const nanoid = customAlphabet('abcdefghijklmnopqrstuvwxyz1234567890', 6);
 
 const nodeTypes = {
   [FlowModuleTypeEnum.userGuide]: NodeUserGuide,
@@ -104,130 +60,16 @@ const edgeTypes = {
 };
 type Props = { app: AppSchema; onCloseSettings: () => void };
 
-const AppEdit = ({ app, onCloseSettings }: Props) => {
+function FlowHeader({ app, onCloseSettings }: Props & {}) {
   const theme = useTheme();
-  const { toast } = useToast();
   const { t } = useTranslation();
   const { copyData } = useCopyData();
-
-  const reactFlowWrapper = useRef<HTMLDivElement>(null);
   const ChatTestRef = useRef<ChatTestComponentRef>(null);
-
-  const { updateAppDetail } = useUserStore();
-  const { x, y, zoom } = useViewport();
-  const [nodes, setNodes, onNodesChange] = useNodesState<FlowModuleItemType>([]);
-  const [edges, setEdges, onEdgesChange] = useEdgesState([]);
-  const {
-    isOpen: isOpenTemplate,
-    onOpen: onOpenTemplate,
-    onClose: onCloseTemplate
-  } = useDisclosure();
   const { isOpen: isOpenImport, onOpen: onOpenImport, onClose: onCloseImport } = useDisclosure();
+  const { updateAppDetail } = useUserStore();
+  const { nodes, edges, onFixView } = useFlowStore();
 
   const [testModules, setTestModules] = useState<AppModuleItemType[]>();
-
-  const onFixView = useCallback(() => {
-    const btn = document.querySelector('.react-flow__controls-fitview') as HTMLButtonElement;
-
-    setTimeout(() => {
-      btn && btn.click();
-    }, 100);
-  }, []);
-
-  const onAddNode = useCallback(
-    ({ template, position }: { template: FlowModuleTemplateType; position: XYPosition }) => {
-      if (!reactFlowWrapper.current) return;
-      const reactFlowBounds = reactFlowWrapper.current.getBoundingClientRect();
-      const mouseX = (position.x - reactFlowBounds.left - x) / zoom - 100;
-      const mouseY = (position.y - reactFlowBounds.top - y) / zoom;
-
-      setNodes((state) =>
-        state.concat(
-          appModule2FlowNode({
-            item: {
-              ...template,
-              moduleId: nanoid(),
-              position: { x: mouseX, y: mouseY }
-            },
-            onChangeNode,
-            onDelNode,
-            onDelEdge,
-            onCopyNode,
-            onCollectionNode
-          })
-        )
-      );
-    },
-    [x, zoom, y]
-  );
-  const onDelNode = useCallback(
-    (nodeId: string) => {
-      setNodes((state) => state.filter((item) => item.id !== nodeId));
-      setEdges((state) => state.filter((edge) => edge.source !== nodeId && edge.target !== nodeId));
-    },
-    [setEdges, setNodes]
-  );
-  const onDelEdge = useCallback(
-    ({
-      moduleId,
-      sourceHandle,
-      targetHandle
-    }: {
-      moduleId: string;
-      sourceHandle?: string;
-      targetHandle?: string;
-    }) => {
-      if (!sourceHandle && !targetHandle) return;
-      setEdges((state) =>
-        state.filter((edge) => {
-          if (edge.source === moduleId && edge.sourceHandle === sourceHandle) return false;
-          if (edge.target === moduleId && edge.targetHandle === targetHandle) return false;
-
-          return true;
-        })
-      );
-    },
-    [setEdges]
-  );
-  const onCopyNode = useCallback(
-    (nodeId: string) => {
-      setNodes((nodes) => {
-        const node = nodes.find((node) => node.id === nodeId);
-        if (!node) return nodes;
-        const template = {
-          logo: node.data.logo,
-          name: node.data.name,
-          intro: node.data.intro,
-          description: node.data.description,
-          flowType: node.data.flowType,
-          inputs: node.data.inputs,
-          outputs: node.data.outputs,
-          showStatus: node.data.showStatus
-        };
-        return nodes.concat(
-          appModule2FlowNode({
-            item: {
-              ...template,
-              moduleId: nanoid(),
-              position: { x: node.position.x + 200, y: node.position.y + 50 }
-            },
-            onChangeNode,
-            onDelNode,
-            onDelEdge,
-            onCopyNode,
-            onCollectionNode
-          })
-        );
-      });
-    },
-    [setNodes]
-  );
-  const onCollectionNode = useCallback(
-    (nodeId: string) => {
-      console.log(nodes.find((node) => node.id === nodeId));
-    },
-    [nodes]
-  );
 
   const flow2AppModules = useCallback(() => {
     const modules: AppModuleItemType[] = nodes.map((item) => ({
@@ -271,106 +113,6 @@ const AppEdit = ({ app, onCloseSettings }: Props) => {
     });
     return modules;
   }, [edges, nodes]);
-  const onChangeNode = useCallback(
-    ({ moduleId, key, type = 'inputs', value }: FlowModuleItemChangeProps) => {
-      setNodes((nodes) =>
-        nodes.map((node) => {
-          if (node.id !== moduleId) return node;
-
-          const updateObj: Record<string, any> = {};
-
-          if (type === 'inputs') {
-            updateObj.inputs = node.data.inputs.map((item) => (item.key === key ? value : item));
-          } else if (type === 'addInput') {
-            const input = node.data.inputs.find((input) => input.key === value.key);
-            if (input) {
-              toast({
-                status: 'warning',
-                title: 'key 重复'
-              });
-              updateObj.inputs = node.data.inputs;
-            } else {
-              updateObj.inputs = node.data.inputs.concat(value);
-            }
-          } else if (type === 'delInput') {
-            onDelEdge({ moduleId, targetHandle: key });
-            updateObj.inputs = node.data.inputs.filter((item) => item.key !== key);
-          } else if (type === 'attr') {
-            updateObj[key] = value;
-          } else if (type === 'outputs') {
-            // del output connect
-            const delOutputs = node.data.outputs.filter(
-              (item) => !value.find((output: FlowOutputTargetItemType) => output.key === item.key)
-            );
-            delOutputs.forEach((output) => {
-              onDelEdge({ moduleId, sourceHandle: output.key });
-            });
-            updateObj.outputs = value;
-          }
-
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              ...updateObj
-            }
-          };
-        })
-      );
-    },
-    []
-  );
-
-  const onDelConnect = useCallback((id: string) => {
-    setEdges((state) => state.filter((item) => item.id !== id));
-  }, []);
-  const onConnect = useCallback(
-    ({ connect }: { connect: Connection }) => {
-      const source = nodes.find((node) => node.id === connect.source)?.data;
-      const sourceType = (() => {
-        if (source?.flowType === FlowModuleTypeEnum.classifyQuestion) {
-          return FlowValueTypeEnum.boolean;
-        }
-        return source?.outputs.find((output) => output.key === connect.sourceHandle)?.valueType;
-      })();
-
-      const targetType = nodes
-        .find((node) => node.id === connect.target)
-        ?.data?.inputs.find((input) => input.key === connect.targetHandle)?.valueType;
-
-      if (!sourceType || !targetType) {
-        return toast({
-          status: 'warning',
-          title: t('app.Connection is invalid')
-        });
-      }
-      if (
-        sourceType !== FlowValueTypeEnum.any &&
-        targetType !== FlowValueTypeEnum.any &&
-        sourceType !== targetType
-      ) {
-        return toast({
-          status: 'warning',
-          title: t('app.Connection type is different')
-        });
-      }
-
-      setEdges((state) =>
-        addEdge(
-          {
-            ...connect,
-            type: 'buttonedge',
-            animated: true,
-            data: {
-              onDelete: onDelConnect
-            }
-          },
-          state
-        )
-      );
-    },
-    [nodes]
-  );
 
   const { mutate: onclickSave, isLoading } = useRequest({
     mutationFn: () => {
@@ -386,49 +128,8 @@ const AppEdit = ({ app, onCloseSettings }: Props) => {
     }
   });
 
-  const initData = useCallback(
-    (modules: AppModuleItemType[]) => {
-      const edges = appModule2FlowEdge({
-        modules,
-        onDelete: onDelConnect
-      });
-      setEdges(edges);
-
-      setNodes(
-        modules.map((item) =>
-          appModule2FlowNode({
-            item,
-            onChangeNode,
-            onDelNode,
-            onDelEdge,
-            onCopyNode,
-            onCollectionNode
-          })
-        )
-      );
-
-      onFixView();
-    },
-    [
-      onDelConnect,
-      setEdges,
-      setNodes,
-      onFixView,
-      onChangeNode,
-      onDelNode,
-      onDelEdge,
-      onCopyNode,
-      onCollectionNode
-    ]
-  );
-
-  useEffect(() => {
-    initData(JSON.parse(JSON.stringify(app.modules)));
-  }, [app.modules]);
-
   return (
     <>
-      {/* header */}
       <Flex
         py={3}
         px={[2, 5, 8]}
@@ -515,6 +216,38 @@ const AppEdit = ({ app, onCloseSettings }: Props) => {
           />
         </MyTooltip>
       </Flex>
+      {isOpenImport && <ImportSettings onClose={onCloseImport} />}
+      <ChatTest
+        ref={ChatTestRef}
+        modules={testModules}
+        app={app}
+        onClose={() => setTestModules(undefined)}
+      />
+    </>
+  );
+}
+const Header = React.memo(FlowHeader);
+
+const AppEdit = (props: Props) => {
+  const { app } = props;
+
+  const {
+    isOpen: isOpenTemplate,
+    onOpen: onOpenTemplate,
+    onClose: onCloseTemplate
+  } = useDisclosure();
+
+  const { reactFlowWrapper, nodes, onNodesChange, edges, onEdgesChange, onConnect, initData } =
+    useFlowStore();
+
+  useEffect(() => {
+    initData(JSON.parse(JSON.stringify(app.modules)));
+  }, [app.modules]);
+
+  return (
+    <>
+      {/* header */}
+      <Header {...props} />
       <Box
         minH={'400px'}
         flex={'1 0 0'}
@@ -571,43 +304,24 @@ const AppEdit = ({ app, onCloseSettings }: Props) => {
           <Controls position={'bottom-right'} style={{ display: 'flex' }} showInteractive={false} />
         </ReactFlow>
 
-        <TemplateList
-          isOpen={isOpenTemplate}
-          nodes={nodes}
-          onAddNode={onAddNode}
-          onClose={onCloseTemplate}
-        />
-        <ChatTest
-          ref={ChatTestRef}
-          modules={testModules}
-          app={app}
-          onClose={() => setTestModules(undefined)}
-        />
+        <TemplateList isOpen={isOpenTemplate} nodes={nodes} onClose={onCloseTemplate} />
       </Box>
-      {isOpenImport && (
-        <ImportSettings
-          onClose={onCloseImport}
-          onSuccess={(data) => {
-            setEdges([]);
-            setNodes([]);
-            setTimeout(() => {
-              initData(data);
-            }, 10);
-          }}
-        />
-      )}
     </>
   );
 };
 
-const Flow = (data: Props) => (
-  <Box h={'100%'} position={'fixed'} zIndex={999} top={0} left={0} right={0} bottom={0}>
-    <ReactFlowProvider>
-      <Flex h={'100%'} flexDirection={'column'} bg={'#fff'}>
-        {!!data.app._id && <AppEdit {...data} />}
-      </Flex>
-    </ReactFlowProvider>
-  </Box>
-);
+const Flow = (data: Props) => {
+  return (
+    <Box h={'100%'} position={'fixed'} zIndex={999} top={0} left={0} right={0} bottom={0}>
+      <ReactFlowProvider>
+        <FlowProvider>
+          <Flex h={'100%'} flexDirection={'column'} bg={'#fff'}>
+            {!!data.app._id && <AppEdit {...data} />}
+          </Flex>
+        </FlowProvider>
+      </ReactFlowProvider>
+    </Box>
+  );
+};
 
 export default React.memo(Flow);
