+++ projects/app/src/service/moduleDispatch/kb/search.ts
@@ -7,8 +7,7 @@ import type { SelectedDatasetType } from '@/types/core/dataset';
 import type { QuoteItemType } from '@/types/chat';
 import { PgDatasetTableName } from '@/constants/plugin';
 import { FlowModuleTypeEnum } from '@/constants/flow';
-import { ModuleDispatchProps } from '@/types/core/modules';
-
+import type { ModuleDispatchProps } from '@/types/core/chat/type';
 type KBSearchProps = ModuleDispatchProps<{
   kbList: SelectedDatasetType;
   similarity: number;
