+++ projects/app/src/service/moduleDispatch/tools/answer.ts
@@ -1,8 +1,7 @@
 import { sseResponseEventEnum, TaskResponseKeyEnum } from '@/constants/chat';
 import { sseResponse } from '@/service/utils/tools';
 import { textAdaptGptResponse } from '@/utils/adapt';
-import type { ModuleDispatchProps } from '@/types/core/modules';
-
+import type { ModuleDispatchProps } from '@/types/core/chat/type';
 export type AnswerProps = ModuleDispatchProps<{
   text: string;
 }>;
